PythonUv/pyproject.toml
[project]
name = "pythonuv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "motor>=3.4.0",
    "python-dotenv>=1.0.1",
]
